Project 3 - Colour Flipper
This project focused more on the JS, both the CSS and HTML had been already coded. Need to take a closer look into the logic behind this project; felt quite rushed when I was doing the code-along. Would have liked to have a bit more input on the CSS aspect of the project. Nevertheless, it did introduce me to some new concepts such as addEventListener. Was good practice for generating random numbers which I had learned from previous projects.