Project 8 - Live User Filter
Update - This is definitely the hardest project I've completed. WIll need to go over this a couple of times to get a proper understanding of the JS involved. Not entirely impressed with the outcome of my project compared to how it is meant to look. Once I get more confident in my JS then I will need to revisit this and correct it properly.  Going to upload it now as it is, more so for my own progress but feel free to look and comment on this with how I can improve it